PORT API CHANGE: uPortGetTickTimeMs() now returns int32_t. (#643)
Changed the function to return int32_t instead of int64_t.
The function could wrap around silently since it was using
int32_t anyway on some platforms (for example STM32Cube and nRF5 SDK).
In addition, changed a few timeout checks to avoid wrap-around in
the calculation.

+++ cell/src/u_cell_private.h
@@ -388,8 +388,8 @@ typedef struct uCellPrivateInstance_t {
     networkStatus[U_CELL_NET_REG_DOMAIN_MAX_NUM]; /**< Registation status in each domain. */
     uCellNetRat_t rat[U_CELL_NET_REG_DOMAIN_MAX_NUM];  /**< The active RAT for each domain. */
     uCellPrivateRadioParameters_t radioParameters; /**< The radio parameters. */
-    int64_t startTimeMs;     /**< Used while connecting and scanning. */
-    int64_t connectedAtMs;   /**< When a connection was last established,
+    int32_t startTimeMs;     /**< Used while connecting and scanning. */
+    int32_t connectedAtMs;   /**< When a connection was last established,
                                   can be used for offsetting from that time;
                                   does NOT mean that we are currently connected. */
     bool rebootIsRequired;   /**< Set to true if a reboot of the module is
