+++ port/platform/cell_ucpu/r5/src/u_port.c
@@ -189,7 +189,7 @@ void uPortDeinit()
 }
 
 // Get the current tick converted to a time in milliseconds.
-int64_t uPortGetTickTimeMs()
+int32_t uPortGetTickTimeMs()
 {
     return tx_time_get();
 }
