+++ port/platform/zephyr/src/u_port.c
@@ -112,7 +112,7 @@ void uPortDeinit()
 }
 
 // Get the current tick converted to a time in milliseconds.
-int64_t uPortGetTickTimeMs()
+int32_t uPortGetTickTimeMs()
 {
     return k_uptime_get();
 }
